requirements.txt
pyyaml
adventurelib
colorama
pyfiglet
git+https://github.com/ccaroon/scriptum.git
invoke
